# Improve vector selector performance