--- a/engine/enginefuzz_test.go
+++ b/engine/enginefuzz_test.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/efficientgo/core/testutil"
 	"github.com/prometheus/prometheus/promql"
+
 	"github.com/thanos-community/promql-engine/engine"
 )