--- a/physicalplan/unary/unary.go
+++ b/physicalplan/unary/unary.go
@@ -1,3 +1,6 @@
+// Copyright (c) The Thanos Community Authors.
+// Licensed under the Apache License 2.0.
+
 package unary
 
 import (